+++ relay/src/Contracts/BillingNoticeTransport.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace AxelFerdinand\StatamicSecretaryRelay\Contracts;
+
+use AxelFerdinand\StatamicSecretaryRelay\Data\BillingNotice;
+
+interface BillingNoticeTransport
+{
+    public function send(BillingNotice $notice): string;
+}